+++ swagger_parser/test/e2e/tests/multipart_request_with_ref_no_properties/expected_files/clients/fallback_client.dart
@@ -1,0 +1,18 @@
+// coverage:ignore-file
+// GENERATED CODE - DO NOT MODIFY BY HAND
+// ignore_for_file: type=lint, unused_import, invalid_annotation_target, unnecessary_import
+
+import 'package:dio/dio.dart';
+import 'package:retrofit/retrofit.dart';
+
+part 'fallback_client.g.dart';
+
+@RestApi()
+abstract class FallbackClient {
+  factory FallbackClient(Dio dio, {String? baseUrl}) = _FallbackClient;
+
+  /// import with an empty multipart body
+  @MultiPart()
+  @POST('/multipart/request/empty-ref')
+  Future<String> postMultipartRequestEmptyRef();
+}